Fix error starting applications when dev tools is on the classpath
I could only reproduce an issue when:

1. devtools on the classpath
2. spring-boot-starter-data-jpa (and an impl) on the classpath
3. okta-spring-boot-starter was NOT available in the same Intellij project

The fix was ultimately making the TokenService class public

Fixes #40
